requirements.txt
# ============================================================================
# Research Agent - Main Dependencies
# ============================================================================
# Agent Runtime, Local Development, and Testing

# ============================================================================
# AWS Bedrock AgentCore
# ============================================================================
bedrock-agentcore>=1.0.4
bedrock-agentcore-starter-toolkit>=0.1.26

# ============================================================================
# LangChain Stack
# ============================================================================
langchain>=0.3.0
langchain-core>=0.3.0
langchain-aws>=0.2.0
langchain-community>=0.3.0
langchain-mcp-adapters>=0.1.0

# ============================================================================
# LangGraph for Agent Orchestration
# ============================================================================
langgraph>=0.6.0
langgraph-checkpoint>=2.0.0
langgraph-checkpoint-aws>=0.2.0
langgraph-prebuilt>=0.6.0

# ============================================================================
# LangSmith (Optional - for tracing and monitoring)
# ============================================================================
langsmith>=0.1.0

# ============================================================================
# AWS SDK
# ============================================================================
boto3>=1.40.0
botocore>=1.40.0

# ============================================================================
# Search and Research Tools
# ============================================================================
ddgs>=9.0.0                    # DuckDuckGo search (renamed from duckduckgo-search)
arxiv>=2.0.0                   # ArXiv paper search
wikipedia-api>=0.8.0           # Wikipedia search and article retrieval
yfinance>=0.2.0                # Yahoo Finance stock data
tavily-python>=0.3.0           # Tavily AI search (optional, for local testing)

# ============================================================================
# Document Processing
# ============================================================================
pymupdf>=1.24.0                # PDF processing
python-docx>=1.1.0             # Word document generation
beautifulsoup4>=4.12.0         # HTML parsing

# ============================================================================
# Data Processing and Visualization
# ============================================================================
pandas>=2.0.0                  # Data manipulation
matplotlib>=3.7.0              # Chart generation
numpy>=1.24.0                  # Numerical computing

# ============================================================================
# Utilities
# ============================================================================
python-dotenv>=1.0.0           # Environment variable management
pyyaml>=6.0.0                  # YAML configuration
pydantic>=2.0.0                # Data validation
nest-asyncio>=1.5.0            # Async support
requests>=2.31.0               # HTTP client

# ============================================================================
# Gateway Integration
# ============================================================================
mcp>=1.1.0                     # MCP protocol implementation
fastmcp>=2.0.0                 # MCP client for Gateway testing (optional)
websockets>=15.0.1             # WebSocket support (required by fastmcp)
httpx>=0.27.0                  # HTTP client for MCP transport
anyio>=4.0.0                   # Async I/O support for MCP

